requirements.txt
amqp==2.5.2
asgiref==3.2.5
astroid==2.3.3
autopep8==1.5
billiard==3.6.3.0
celery==4.4.1
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.4
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
drf-yasg==1.17.1
entrypoints==0.3
flake8==3.7.9
idna==2.9
importlib-metadata==1.5.0
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.1
kombu==4.6.8
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.3
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
pyparsing==2.4.6
pytz==2019.3
redis==3.4.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wrapt==1.11.2
zipp==3.1.0
